JS Stalker Exercise
Asks for user's first name, last name and age. I added a bootstrap CDN and my own CSS to make it look how I wanted. I also had it print out to an alert *and* to the console. In his exercise, he just wants you to print to the console.